refactor(input): Split role all input to woowahanInput
